backend/pyproject.toml
[project]
name = "deepresearch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.3.2",
    "fastapi>=0.120.4",
    "google-adk>=1.18.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.38.0",
]
